requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
black==22.8.0
cachetools==4.2.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tomli==2.0.1
tornado==6.1
typed-ast==1.5.4
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
zipp==3.8.1
